Version bump
We've expanded our peer dependencies to allow React 18.
This isn't really a breaking change, but usage of this with
React >16 is under-tested. So we're going to do a major
version bump so that any projects that depend on this
package don't implicitly pick up this new version,
resolve to a different version of React, and suddenly find
old things not working.

Also, we revamped the build code and there are minor
compiler / TypeScript-ish changes there, so a little
extra caution feels warranted.